CB-5286 - Fix warnings when compiled under arm64

+++ CordovaLib/Classes/CDVShared.m
@@ -50,8 +50,8 @@
 - (NSString*)JSONRepresentation
 {
     return [NSString stringWithFormat:
-           @"{ code: %d, message: '%@'}",
-           self.code,
+           @"{ code: %ld, message: '%@'}",
+           (long)self.code,
            [self localizedDescription]
     ];
 }
