+++ CordovaLib/Classes/CDVWebViewDelegate.m
@@ -139,13 +139,13 @@ static NSString *stripFragment(NSString* url)
 {
     NSString* loadToken = [webView stringByEvaluatingJavaScriptFromString:@"window.__cordovaLoadToken"];
 
-    return [[NSString stringWithFormat:@"%d", _curLoadToken] isEqualToString:loadToken];
+    return [[NSString stringWithFormat:@"%ld", (long)_curLoadToken] isEqualToString:loadToken];
 }
 
 - (void)setLoadToken:(UIWebView*)webView
 {
     _curLoadToken += 1;
-    [webView stringByEvaluatingJavaScriptFromString:[NSString stringWithFormat:@"window.__cordovaLoadToken=%d", _curLoadToken]];
+    [webView stringByEvaluatingJavaScriptFromString:[NSString stringWithFormat:@"window.__cordovaLoadToken=%ld", (long)_curLoadToken]];
 }
 
 - (NSString*)evalForCurrentURL:(UIWebView*)webView
@@ -223,7 +223,7 @@ static NSString *stripFragment(NSString* url)
                     // Redirect case.
                     // We expect loadCount == 1.
                     if (_loadCount != 1) {
-                        NSLog(@"CDVWebViewDelegate: Detected redirect when loadCount=%d", _loadCount);
+                        NSLog(@"CDVWebViewDelegate: Detected redirect when loadCount=%ld", (long)_loadCount);
                     }
                     break;
 
@@ -239,7 +239,7 @@ static NSString *stripFragment(NSString* url)
                     {
                         _loadCount = 0;
                         _state = STATE_WAITING_FOR_LOAD_START;
-                        NSString* description = [NSString stringWithFormat:@"CDVWebViewDelegate: Navigation started when state=%d", _state];
+                        NSString* description = [NSString stringWithFormat:@"CDVWebViewDelegate: Navigation started when state=%ld", (long)_state];
                         NSLog(@"%@", description);
                         if ([_delegate respondsToSelector:@selector(webView:didFailLoadWithError:)]) {
                             NSDictionary* errorDictionary = @{NSLocalizedDescriptionKey : description};
@@ -287,7 +287,7 @@ static NSString *stripFragment(NSString* url)
 
         case STATE_WAITING_FOR_LOAD_START:
             if (_loadCount != 0) {
-                NSLog(@"CDVWebViewDelegate: Unexpected loadCount in didStart. count=%d", _loadCount);
+                NSLog(@"CDVWebViewDelegate: Unexpected loadCount in didStart. count=%ld", (long)_loadCount);
             }
             fireCallback = YES;
             _state = STATE_WAITING_FOR_LOAD_FINISH;
@@ -307,7 +307,7 @@ static NSString *stripFragment(NSString* url)
             break;
 
         default:
-            NSLog(@"CDVWebViewDelegate: Unexpected didStart with state=%d loadCount=%d", _state, _loadCount);
+            NSLog(@"CDVWebViewDelegate: Unexpected didStart with state=%ld loadCount=%ld", (long)_state, (long)_loadCount);
     }
     VerboseLog(@"webView didStartLoad (after). state=%d loadCount=%d fireCallback=%d", _state, _loadCount, fireCallback);
     if (fireCallback && [_delegate respondsToSelector:@selector(webViewDidStartLoad:)]) {
